Replace old VC with new refactored VC
